refactor: Group every component about profile into same folder
