+++ src/components/Profile/PasswordSection.spec.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { screen, render } from '@testing-library/react'
 
-import AppLike from '../../test/AppLike'
-import PasswordSection from 'components/PasswordSection'
+import AppLike from '../../../test/AppLike'
+import PasswordSection from 'components/Profile/PasswordSection'
 import { useHasPassword } from 'hooks/useHasPassword'
 
 jest.mock('hooks/useHasPassword', () => ({
