+++ src/components/Profile/ProfileView.jsx
@@ -7,16 +7,17 @@ import Spinner from 'cozy-ui/transpiled/react/Spinner'
 import Stack from 'cozy-ui/transpiled/react/Stack'
 
 import Page from 'components/Page'
-import DeleteAccount from 'components/DeleteAccount'
-import Input from 'components/Input'
-import LanguageSection from 'components/LanguageSection'
-import DefaultRedirectionSection from 'components/DefaultRedirectionSection/DefaultRedirectionSection'
-import TwoFA from 'components/2FA'
-import Import from 'components/Import'
-import ExportSection from 'components/export/ExportSection'
-import TrackingSection from 'components/TrackingSection'
 import PageTitle from 'components/PageTitle'
-import PasswordSection from 'components/PasswordSection'
+import Input from 'components/Input'
+
+import DeleteAccount from 'components/Profile/DeleteAccount'
+import LanguageSection from 'components/Profile/LanguageSection'
+import DefaultRedirectionSection from 'components/Profile/DefaultRedirectionSection'
+import TwoFA from 'components/2FA'
+import Import from 'components/Profile/Import'
+import ExportSection from 'components/export/ExportSection'
+import TrackingSection from 'components/Profile/TrackingSection'
+import PasswordSection from 'components/Profile/PasswordSection'
 import EmailSection from 'components/Email/EmailSection'
 
 const ProfileView = ({
